+++ packages/adminui-vue/components/UbEntityEditComponent.vue
@@ -8,12 +8,11 @@
     @keyup.alt.r="remove"
   >
     <toolbar
-      v-if="value.ID"
-      v-model="value.ID"
+      :instance-id="instance.ID"
       :entity-name="entityName"
       :is-new="isNew"
       :is-changed="isChanged"
-      :simple-audit="{mi_createDate: value.mi_createDate, mi_modifyDate: value.mi_modifyDate}"
+      :simple-audit="{mi_createDate: instance.mi_createDate, mi_modifyDate: instance.mi_modifyDate}"
       :use-only-own-actions="useOnlyOwnActions"
       :input-actions="inputActions"
       :input-buttons="inputButtons"
@@ -29,18 +28,22 @@
 </template>
 
 <script>
+// TODO rename UbToolbarComponent to UFormToolbar
 const toolbar = require('./UbToolbarComponent.vue').default
 
 module.exports = {
   name: 'UbEntityEditComponent',
   props: {
-    value: Object,
+    instance: {
+      type: Object,
+      required: true
+    },
     entityName: {
       type: String,
       required: true
     },
     externalData: Object,
-    instanceID: Number,
+    instanceId: Number,
     useOnlyOwnActions: Boolean,
     inputActions: Array,
     inputButtons: Array,
@@ -52,7 +55,7 @@ module.exports = {
     return {
       loading: false,
       isNew: false,
-      oldData: Object.assign({}, this.value)
+      originalData: Object.assign({}, this.instance) // TODO deep clone
     }
   },
   computed: {
@@ -72,8 +75,8 @@ module.exports = {
     },
     changedColumns () {
       let result = {}
-      Object.keys(this.value).forEach(field => {
-        if (this.value[field] !== this.oldData[field]) result[field] = this.value[field]
+      Object.keys(this.instance).forEach(field => {
+        if (this.instance[field] !== this.originalData[field]) result[field] = this.instance[field]
       })
       return result
     },
@@ -88,8 +91,8 @@ module.exports = {
         data.resultData.data[0].forEach((item, index) => {
           object[data.resultData.fields[index]] = item
         })
-        this.$emit('input', object)
-        this.oldData = Object.assign({}, object)
+        this.$emit('data-loaded', object)
+        this.originalData = Object.assign({}, object)
       })
     },
     saveAndClose () {
@@ -100,7 +103,7 @@ module.exports = {
     },
     saveEntity (callback) {
       let changedColumns = Object.assign({}, this.changedColumns)
-      if (Object.keys(changedColumns).length > 0) {
+      if (Object.keys(changedColumns).length) {
         let saveFn = () => {
           Object.keys(changedColumns).forEach(col => {
             if (this.entitySchema.attributes[col] && this.entitySchema.attributes[col].isMultiLang && Object.keys(changedColumns).some(key => key.indexOf(`${col}_`) !== -1)) {
@@ -108,8 +111,10 @@ module.exports = {
               delete changedColumns[col]
             }
           })
-          changedColumns.ID = this.value.ID
-          changedColumns.mi_modifyDate = this.value.mi_modifyDate
+          changedColumns.ID = this.instance.ID
+          if (this.instance.mi_modifyDate) {
+            changedColumns.mi_modifyDate = this.instance.mi_modifyDate
+          }
           let params = {
             fieldList: this.entityFields,
             entity: this.entitySchema.name,
@@ -139,61 +144,65 @@ module.exports = {
     },
     remove () {
       this.loading = true
-      this.$dialogYesNo('deletionDialogConfirmCaption', this.$ut('deleteFormConfirmCaption', this.value[this.entitySchema.descriptionAttribute]))
-        .then(res => {
-          if (!res) {
-            this.loading = false
-            return
-          }
-          let request = {
-            entity: this.entitySchema.name,
-            method: 'delete',
-            execParams: {
-              ID: this.value.ID
-            }
-          }
-          return this.$UB.connection.doDelete(request).then(result => {
-            this.$UB.connection.emit(`${this.entitySchema.name}:changed`, result.execParams.ID)
-            this.$UB.connection.emit(`${this.entitySchema.name}:delete`, result.execParams.ID)
-            this.currentTab.close()
-          }).finally(() => {
-            this.loading = false
-          })
-        })
-    }
-  },
-  mounted () {
-    if (!this.value || Object.keys(this.value) < 1) {
-      let dataP
-      this.loading = true
-      if (this.instanceID) {
-        dataP = this.$UB.Repository(this.entityName).attrs(this.entityFields).selectById(this.instanceID).then(resp => {
-          this.$emit('input', resp)
-          this.oldData = Object.assign({}, resp)
-        })
-      } else {
-        let parameters = {
-          entity: this.entityName,
-          fieldList: this.entityFields
+      this.$dialogYesNo('deletionDialogConfirmCaption',
+        this.$ut('deleteFormConfirmCaption', this.instance[this.entitySchema.descriptionAttribute])
+      ).then(res => {
+        if (!res) {
+          this.loading = false
+          return
         }
-        if (this.externalData) {
-          parameters.execParams = this.externalData
+        let request = {
+          entity: this.entitySchema.name,
+          method: 'delete',
+          execParams: {
+            ID: this.instance.ID
+          }
         }
-        dataP = this.$UB.connection.addNew(parameters).then(result => {
-          let data = {}
-          result.resultData.fields.forEach((item, key) => {
-            data[item] = result.resultData.data[0][key]
-          })
-          this.$emit('input', data)
-          this.oldData = {}
+        return this.$UB.connection.doDelete(request).then(result => {
+          this.$UB.connection.emit(`${this.entitySchema.name}:changed`, result.execParams.ID)
+          this.$UB.connection.emit(`${this.entitySchema.name}:delete`, result.execParams.ID)
+          this.currentTab.close()
+        }).finally(() => {
+          this.loading = false
         })
-        this.isNew = true
-      }
-      dataP.finally(() => {
-        this.$emit('dataLoaded')
-        this.loading = false
       })
     }
+  },
+  /** start loading form data from remote in case it's not passed */
+  beforeMount () {
+    if (this.instance && Object.keys(this.instance).length) return // instance provided by parent
+    let dataP
+    this.loading = true
+    if (this.instanceId) { // existed record
+      dataP = this.$UB.Repository(this.entityName).attrs(this.entityFields).selectById(this.instanceId)
+    } else { // create new
+      let parameters = {
+        entity: this.entityName,
+        fieldList: this.entityFields
+      }
+      if (this.externalData) {
+        parameters.execParams = this.externalData
+      }
+      dataP = this.$UB.connection.addNew(parameters).then(result => {
+        let data = {}
+        result.resultData.fields.forEach((item, key) => {
+          data[item] = result.resultData.data[0][key]
+        })
+        return data
+      })
+      this.isNew = true
+    }
+    dataP.then(resp => {
+      this.originalData = Object.assign({}, resp)
+      // this.instance = resp
+      // this.$emit('input', resp)
+      this.$emit('data-loaded', resp)
+    }).finally(() => {
+      this.loading = false
+    })
+  },
+
+  mounted () {
     this.currentTab.on('beforeClose', function () {
       if (!this.currentTab.forceClose && Object.keys(this.changedColumns).length > 0) {
         $App.viewport.centralPanel.setActiveTab(this.currentTab)