+++ packages/adminui-vue/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  
 ## [1.3.3]
 ### Changed
+ - ToolbarComponent property `value` renamed to `instanceId`. To be used as `<u-form-toolbar :instance-id="formData.ID"...>`
  - prevent override Vue.prototype.$notify introduced by El, instead inject UDialog as `$dialog`
  - Vue forms will use `Alt+R` for "Remove" instead  of Alt+Del because DEl and Backspace because
  on Mac keyboards the key normally identified as "backspace" on PC keyboards is called "delete"